More scripting magic!
Now the Java application calls CommitToGit.py, passing it the current directory and the new weigh in.

The current directory argument is used to allow the project being moved elsewhere in the system (/ making it export-proof), and the new current weight is not actually used by CommitToGit.py, but at its completion, CommitToGit.py will pass that variable on to the next script; UpdateWebsite.py.

UpdateWebsite.py will open Google Chrome and navigate to the admin panel of x4iiiis.com, sign in, and add a new entry to the weigh-in database.

Obviously, I have removed my login credentials before making this commit, and will be adding the file to .gitignore so that CommitToGit.py does not upload my username and password to GitHub!